requirements.txt
pillow
ultralytics
flask
flask_cors
sqlite
opencv-contrib-python
```
